// Client/src/app/model/AppItem.ts
/* spiritdev Softwareentwicklung GmbH 2020
      Author Vladimir Kogan
*/
export class AppItem {

    id: number;
    title: string;
    annotation: string;
    price: number;
    image: string;
    category: string;
    plattform: string;
    projectStart: Date;
    projectEnd: Date;
    salesAmount: number;

}